(thickfold only): Fix unsigned integer overflow caught by the LLVM sanitizer
Conspicuous note:
- regnexec.c: _matchpush(): In this function the sh_match tests will
  fail if we force a size >0, so allow multiplication by 0.

+++ src/lib/libast/regex/regcache.c
@@ -58,9 +58,9 @@ static State_t	matchstate;
 static void
 flushcache(void)
 {
-	size_t		i;
+	ptrdiff_t	i;
 
-	for (i = matchstate.size; i--;)
+	for (i = (ptrdiff_t)matchstate.size; i--;)
 		if (matchstate.cache[i] && matchstate.cache[i]->keep)
 		{
 			matchstate.cache[i]->keep = 0;
